Fix missing type conversion before bitwise negation operation (a cherry-pick of commit 3bd1713cb from 'master')
* backgraph.c (add_edge, reset_back_edge, backwards_height,
update_max_height): Cast to word before "~" operation.
* checksums.c [MPROTECT_VDB) && !DARWIN] (GC_record_fault): Likewise.
* checksums.c (GC_was_faulted, GC_add_block): Likewise.
* dyn_load.c [HAVE_DL_ITERATE_PHDR && PT_GNU_RELRO]
(GC_register_dynlib_callback): Likewise.
* include/private/dbg_mlc.h [(KEEP_BACK_PTRS || MAKE_BACK_GRAPH)
&& ALIGNMENT==1] (HIDE_BACK_PTR): Likewise.
* include/private/gc_priv.h [!NOT_GCBUILD] (GC_unmap_end): Likewise.
* include/private/gcconfig.h [M68K && LINUX && !__ELF__ || I386
&& (SEQUENT || HAIKU || LINUX && !__ELF__ || DJGPP) || MIPS && EWS4800
&& (!_MIPS_SZPTR || _MIPS_SZPTR!=64) || M88K && CX_UX] (DATASTART):
Likewise.
* include/private/gcconfig.h [MIPS && EWS4800 && (!_MIPS_SZPTR
|| _MIPS_SZPTR!=64)] (DATASTART2): Likewise.
* include/private/gcconfig.h [ALPHA && OSF1] (HEURISTIC2_LIMIT):
Likewise.
* malloc.c (GC_extend_size_map): Likewise.
* mark.c (GC_mark_from, GC_push_all, GC_push_all_eager,
GC_push_all_stack): Likewise.
* misc.c (GC_clear_stack, GC_base): Likewise.
* os_dep.c (GC_get_writable_length, GC_get_stack_base,
GC_find_limit_with_bound, GC_get_register_stack_base,
GC_get_main_stack_base, GC_least_described_address,
GC_unix_mmap_get_mem, GC_unmap_start, GC_write_fault_handler,
pagemap_buffered_read, soft_set_grungy_pages, GC_remove_protection,
catch_exception_raise): Likewise.
* ptr_chck.c (GC_is_visible): Likewise.
* reclaim.c (GC_print_block_descr): Likewise.
* darwin_stop_world.c [!DARWIN_DONT_PARSE_STACK] (GC_FindTopOfStack):
Define and use maskedLR local variable (to avoid code duplication);
change ~0x3 to ~0x3UL.
* cord/cordxtra.c (LINE_START): Cast to size_t before "~" operation.
* include/private/gc_priv.h (ROUNDUP_GRANULE_SIZE): Likewise.
* typd_mlc.c (GC_double_descr): Cast to GC_descr before "~" operation.

+++ cord/cordxtra.c
@@ -518,7 +518,7 @@ typedef struct {
 # define DIV_CACHE_SZ(n) ((n) >> LOG_CACHE_SZ)
 # define MOD_LINE_SZ(n) ((n) & (LINE_SZ - 1))
 # define DIV_LINE_SZ(n) ((n) >> LOG_LINE_SZ)
-# define LINE_START(n) ((n) & ~(LINE_SZ - 1))
+# define LINE_START(n) ((n) & ~(size_t)(LINE_SZ - 1))
 
 typedef struct {
     lf_state * state;
